+++ iot_link/network/lwm2m/lwm2m_al/lwm2m_al.c
@@ -53,7 +53,7 @@ static lwm2m_al_op_cb_t s_lwm2m_al_op_cb;
 /////////////////CREATE THE API FOR THE LWM2M LIB////////////////////////////////
 int lwm2m_al_config(void **handle, lwm2m_al_init_param_t *init_param)
 {
-    int ret = LWM2M_ERR;
+    int ret = (int)LWM2M_ERR;
 
     if ((NULL != handle)
         && (NULL != init_param)
@@ -68,7 +68,7 @@ int lwm2m_al_config(void **handle, lwm2m_al_init_param_t *init_param)
 
 int lwm2m_al_deconfig(void *handle)
 {
-    int ret = LWM2M_ERR;
+    int ret = (int)LWM2M_ERR;
 
     if ((NULL != handle)
         && (NULL != s_lwm2m_al_op_cb.ops)
@@ -82,7 +82,7 @@ int lwm2m_al_deconfig(void *handle)
 
 int lwm2m_al_add_object(void *handle, int object_id, int object_instance_id, int resource_id, void *param)
 {
-    int ret = LWM2M_ERR;
+    int ret = (int)LWM2M_ERR;
 
     if ((NULL != handle)
         && (NULL != s_lwm2m_al_op_cb.ops)
@@ -96,7 +96,7 @@ int lwm2m_al_add_object(void *handle, int object_id, int object_instance_id, int
 
 int lwm2m_al_delete_object(void *handle, int object_id)
 {
-    int ret = LWM2M_ERR;
+    int ret = (int)LWM2M_ERR;
 
     if ((NULL != handle)
         && (NULL != s_lwm2m_al_op_cb.ops)
@@ -110,7 +110,7 @@ int lwm2m_al_delete_object(void *handle, int object_id)
 
 int lwm2m_al_connect(void *handle)
 {
-    int ret = LWM2M_ERR;
+    int ret = (int)LWM2M_ERR;
 
     if ((NULL != handle)
         && (NULL != s_lwm2m_al_op_cb.ops)
@@ -124,7 +124,7 @@ int lwm2m_al_connect(void *handle)
 
 int  lwm2m_al_disconnect(void *handle)
 {
-    int ret = LWM2M_ERR;
+    int ret = (int)LWM2M_ERR;
 
     if ((NULL != handle)
         && (NULL != s_lwm2m_al_op_cb.ops)
@@ -138,7 +138,7 @@ int  lwm2m_al_disconnect(void *handle)
 
 int  lwm2m_al_send(void *handle, lwm2m_al_send_param_t *send_param)
 {
-    int ret = LWM2M_ERR;
+    int ret = (int)LWM2M_ERR;
 
     if ((NULL != handle)
         && (NULL != send_param)
@@ -153,7 +153,7 @@ int  lwm2m_al_send(void *handle, lwm2m_al_send_param_t *send_param)
 
 int lwm2m_al_install(lwm2m_al_op_t *op)
 {
-    int ret = LWM2M_ERR;
+    int ret = (int)LWM2M_ERR;
 
     if ((NULL != op) && (NULL == s_lwm2m_al_op_cb.ops))
     {
@@ -167,7 +167,7 @@ int lwm2m_al_install(lwm2m_al_op_t *op)
 
 int lwm2m_al_uninstall(void)
 {
-    int ret = LWM2M_ERR;
+    int ret = (int)LWM2M_ERR;
 
     if (NULL != s_lwm2m_al_op_cb.ops)
     {
