+++ iot_link/oc/oc_lwm2m/oc_lwm2m_al/oc_lwm2m_al.c
@@ -89,7 +89,7 @@ int oc_lwm2m_unregister(const char *name)
 //////////////////////////APPLICATION INTERFACE/////////////////////////////////
 int oc_lwm2m_report(char  *buf, int len, int timeout)
 {
-    int ret = en_oc_lwm2m_err_system;
+    int ret = (int)en_oc_lwm2m_err_system;
 
     if ((NULL != s_oc_lwm2m_ops.opt) && (NULL != s_oc_lwm2m_ops.opt->report))
     {
@@ -102,7 +102,7 @@ int oc_lwm2m_report(char  *buf, int len, int timeout)
 
 int oc_lwm2m_config( oc_config_param_t *param)
 {
-    int ret = en_oc_lwm2m_err_system;
+    int ret = (int)en_oc_lwm2m_err_system;
 
     if ((NULL != s_oc_lwm2m_ops.opt) && (NULL != s_oc_lwm2m_ops.opt->config))
     {
@@ -112,7 +112,7 @@ int oc_lwm2m_config( oc_config_param_t *param)
         }
         else
         {
-            ret = en_oc_lwm2m_err_parafmt;
+            ret = (int)en_oc_lwm2m_err_parafmt;
         }
     }
 
@@ -122,7 +122,7 @@ int oc_lwm2m_config( oc_config_param_t *param)
 
 int oc_lwm2m_deconfig(void)
 {
-    int ret = en_oc_lwm2m_err_system;
+    int ret = (int)en_oc_lwm2m_err_system;
 
     if ((NULL != s_oc_lwm2m_ops.opt) && (NULL != s_oc_lwm2m_ops.opt->deconfig))
     {
